refactor horizons api parser functions
